Clean up and documentation
