+++ Assets/Scripts/Events/EventRegistry.cs
@@ -1,16 +1,25 @@
 ﻿using UnityEngine.Events;
 using UnityEngine;
 
+
 /// <summary>
 /// A class to orchestrate all events
 /// </summary>
 public static class EventRegistry
 {
-    private readonly static TDEventContainer<string> container = new TDEventContainer<string>();
+    private readonly static GenericMap<string> container = new GenericMap<string>();
 
+    #region events
+    /// <summary>
+    /// Register an event with no parameters.
+    /// </summary>
+    /// <param name="name"></param>
+    /// <param name="unityEvent"></param>
+    /// <returns></returns>
     public static bool RegisterEvent(string name, TDEvent unityEvent)
     {
-        if (!container.ContainsKey(name)) {
+        if (!container.ContainsKey(name))
+        {
             container.Add(name, unityEvent);
             return true;
         }
@@ -20,10 +29,17 @@ public static class EventRegistry
         }
     }
 
-
+    /// <summary>
+    /// Register an event with one parameter
+    /// </summary>
+    /// <param name="name"></param>
+    /// <param name="unityEvent"></param>
+    /// <typeparam name="TParam"></typeparam>
+    /// <returns></returns>
     public static bool RegisterEvent<TParam>(string name, TDEvent<TParam> unityEvent)
     {
-        if (!container.ContainsKey(name)) {
+        if (!container.ContainsKey(name))
+        {
             container.Add(name, unityEvent);
             return true;
         }
@@ -33,9 +49,18 @@ public static class EventRegistry
         }
     }
 
+    /// <summary>
+    /// Register an event with two parameters
+    /// </summary>
+    /// <param name="name"></param>
+    /// <param name="unityEvent"></param>
+    /// <typeparam name="TParam1"></typeparam>
+    /// <typeparam name="TParam2"></typeparam>
+    /// <returns></returns>
     public static bool RegisterEvent<TParam1, TParam2>(string name, TDEvent<TParam1, TParam2> unityEvent)
     {
-        if (!container.ContainsKey(name)) {
+        if (!container.ContainsKey(name))
+        {
             container.Add(name, unityEvent);
             return true;
         }
@@ -44,7 +69,16 @@ public static class EventRegistry
             return false;
         }
     }
+    #endregion
 
+
+    #region actions
+    /// <summary>
+    /// Register an action with no parameters
+    /// </summary>
+    /// <param name="name"></param>
+    /// <param name="callback"></param>
+    /// <returns></returns>
     public static bool RegisterAction(string name, UnityAction callback)
     {
         if (container.TryGetValue(name, out TDEvent value))
@@ -59,6 +93,13 @@ public static class EventRegistry
         }
     }
 
+    /// <summary>
+    /// Register an action with one parameter
+    /// </summary>
+    /// <param name="name"></param>
+    /// <param name="callback"></param>
+    /// <typeparam name="TParam"></typeparam>
+    /// <returns></returns>
     public static bool RegisterAction<TParam>(string name, UnityAction<TParam> callback)
     {
         if (container.TryGetValue(name, out TDEvent<TParam> value))
@@ -73,6 +114,14 @@ public static class EventRegistry
         }
     }
 
+    /// <summary>
+    /// Register an action with two parameters
+    /// </summary>
+    /// <param name="name"></param>
+    /// <param name="callback"></param>
+    /// <typeparam name="TParam1"></typeparam>
+    /// <typeparam name="TParam2"></typeparam>
+    /// <returns></returns>
     public static bool RegisterAction<TParam1, TParam2>(string name, UnityAction<TParam1, TParam2> callback)
     {
         if (container.TryGetValue(name, out TDEvent<TParam1, TParam2> value))
@@ -86,4 +135,6 @@ public static class EventRegistry
             return false;
         }
     }
+    #endregion
+
 }
